Error Analysis
The main errors have to do with comparisons between a pointer and an int, or incorrect conversions of either. I believe this is due to the new declarations we made from the code review. I'd recommend you try and see for yourself. Start with the list which is the root of all the other data structures, maybe that will fix all of them.